Handle nullable columns for databases using pointers
Some databases (e.g. clickhouse) is using pointer values to represent
nullable values when querying from the database. The current logic to
transform values to strings when querying data, does not handle this
correctly. It will simply transform the pointer to a string, which
results in the pointer address being printed.

To handle this, the value reported from the sql driver will now tried to
be extracted, if it is a pointer. This means the value of the pointer
will then be transformed to a string, instead of the pointer directly.

Additionally, some cases where this was causing issues in the clickhouse
implementation have been fixed.

Fixes #151